+++ src/features/editor/extensions/nodes/code-block-container/code-block/tests/render.spec.ts
@@ -4,11 +4,11 @@ import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { Editor } from "@tiptap/react";
 import { describe, expect, it } from "vitest";
-import { CodeBlockFileName } from "../code-block-file-name";
-import { DiffCodeBlock } from "../diff-code-block";
-import { DiffCodeLine } from "../diff-code-block/diff-code-line";
-import { CodeBlockContainer } from "../index";
-import { CodeBlock } from "./index";
+import { CodeBlockFileName } from "../../code-block-file-name";
+import { DiffCodeBlock } from "../../diff-code-block";
+import { DiffCodeLine } from "../../diff-code-block/diff-code-line";
+import { CodeBlockContainer } from "../../index";
+import { CodeBlock } from "../index";
 
 describe("HTMLのレンダリング", () => {
   it("JavaScriptコードブロックが正しいHTMLでレンダリングされる", () => {
